Pool balance added

+++ truffle/contracts/Token.sol
@@ -1,17 +1,14 @@
 // SPDX-License-Identifier: MIT
 
 pragma solidity >=0.8.14;
-import "@openzeppelin/contracts/token/ERC20/ERC20.sol"; 
+import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
-contract DevToken is ERC20{
-
+contract DevToken is ERC20 {
     mapping(address => uint256) balances;
 
-    constructor() ERC20("DevToken", "DevToken") {
-    }
+    constructor() ERC20("DevToken", "DevToken") {}
 
-    function claim(address claimer, uint amount) public {
+    function claim(address claimer, uint256 amount) public {
         _mint(claimer, amount);
     }
-    
 }
